% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_data.R
\docType{data}
\name{dm}
\alias{dm}
\title{DM data for Analysis}
\format{
A \code{tibble} with 33 column
\describe{
\item{studyid}{ Study Id }
\item{domain}{ Domain}
\item{usubjid}{ Unique subjid}
\item{aeseq}{ AE Seq}
\item{aespid}{ aespid}
\item{aeterm}{ AE Unique TERM }
\item{aellt}{ AE Low level term}
\item{aelltcd}{ AE Low level Term CD}
}
}
\usage{
dm
}
\description{
A simple \code{tibble}
}
\keyword{datasets}
