% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{My Hello world}
\usage{
hello(x)
}
\arguments{
\item{x}{The name of the person}
}
\value{
The Output from the \code{link{print}}
}
\description{
My Hello world
}
\examples{
hello("Ram")
\dontrun{
hello("Ram")
}
}
