% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_data.R
\docType{data}
\name{s_data}
\alias{s_data}
\title{Sample data for Analysis}
\format{
A \code{tibble} with 2 column , which are:
\describe{
\item{x}{ random normal data }
\item{y}{ Outcome that is normal based on \code{x}}
}
}
\usage{
s_data
}
\description{
A simple \code{tibble} that is for simple testing and running of examples
}
\keyword{datasets}
