% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_read.R
\name{s_read}
\alias{s_read}
\title{Read CSV duplicate}
\usage{
s_read(path)
}
\arguments{
\item{path}{to file name}
}
\value{
\code{tibble}
}
\description{
Read CSV duplicate
}
\examples{
csv=file.path("c:","users","tatav","OneDrive", "Desktop" , "sas2r","clinicalfd","data-raw","inst","extdata", "s_data.csv")
s_read(csv)
}
