% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_read.R, R/cm_read.R, R/dm_read.R,
%   R/ex_read.R, R/lbch_read.R, R/lbhe_read.R, R/mh_read.R, R/qsco_read.R,
%   R/qsda_read.R, R/qsgi_read.R, R/s_read.R
\name{s_read}
\alias{s_read}
\title{Read CSV duplicate}
\usage{
s_read(path)

s_read(path)

s_read(path)

s_read(path)

s_read(path)

s_read(path)

s_read(path)

s_read(path)

s_read(path)

s_read(path)

s_read(path)
}
\arguments{
\item{path}{to file name}
}
\value{
\code{tibble}

\code{tibble}

\code{tibble}

\code{tibble}

\code{tibble}

\code{tibble}

\code{tibble}

\code{tibble}

\code{tibble}

\code{tibble}

\code{tibble}
}
\description{
Read CSV duplicate

Read CSV duplicate

Read CSV duplicate

Read CSV duplicate

Read CSV duplicate

Read CSV duplicate

Read CSV duplicate

Read CSV duplicate

Read CSV duplicate

Read CSV duplicate

Read CSV duplicate
}
\examples{
csv=file.path("c:","users","tatav","OneDrive", "Desktop" , "sas2r","clinicalfd","data-raw","inst","extdata", "ae.csv")
s_read(csv)
csv=file.path("c:","users","tatav","OneDrive", "Desktop" , "sas2r","clinicalfd","data-raw","inst","extdata", "cm.csv")
s_read(csv)
csv=file.path("c:","users","tatav","OneDrive", "Desktop" , "sas2r","clinicalfd","data-raw","inst","extdata", "dm.csv")
s_read(csv)
csv=file.path("c:","users","tatav","OneDrive", "Desktop" , "sas2r","clinicalfd","data-raw","inst","extdata", "ex.csv")
s_read(csv)
csv=file.path("c:","users","tatav","OneDrive", "Desktop" , "sas2r","clinicalfd","data-raw","inst","extdata", "lbch.csv")
s_read(csv)
csv=file.path("c:","users","tatav","OneDrive", "Desktop" , "sas2r","clinicalfd","data-raw","inst","extdata", "lbhe.csv")
s_read(csv)
csv=file.path("c:","users","tatav","OneDrive", "Desktop" , "sas2r","clinicalfd","data-raw","inst","extdata", "mh.csv")
s_read(csv)
csv=file.path("c:","users","tatav","OneDrive", "Desktop" , "sas2r","clinicalfd","data-raw","inst","extdata", "qsco.csv")
s_read(csv)
csv=file.path("c:","users","tatav","OneDrive", "Desktop" , "sas2r","clinicalfd","data-raw","inst","extdata", "qsda.csv")
s_read(csv)
csv=file.path("c:","users","tatav","OneDrive", "Desktop" , "sas2r","clinicalfd","data-raw","inst","extdata", "qsgi.csv")
s_read(csv)
csv=file.path("c:","users","tatav","OneDrive", "Desktop" , "sas2r","clinicalfd","data-raw","inst","extdata", "s_data.csv")
s_read(csv)
}
